Added the concept of AbstractOfficeTask and TextReplaceOfficeTask.
It is a new use case. It is possible to request tasks over an already
opened document. Modified the OfficeDocumentUtils to provide a static
method to load a document. Added a method so that the OfficeManagers can
return the OfficeContext, and so on you can request the Managers to
execute Document Tasks over a Document that is already opened. As
example, I provide the TextReplace JUnit test. Also, this model can be
extended to provide tasks for every Text Operation, such as find, apply
Style...